doDelete, doModify 권한체크. ResultData에 data1Name 추가

+++ src/main/java/com/khw/exam/demo/controller/UsrArticleController.java
@@ -47,7 +47,7 @@ public class UsrArticleController {
 		int id =(int)writeArticleRd.getData1();
 		Article article = articleService.getArticle(id);
 		
-		return ResultData.from(writeArticleRd.getResultCode(),writeArticleRd.getMsg(),article);
+		return ResultData.from(writeArticleRd.getResultCode(),writeArticleRd.getMsg(),"article",article);
 	}
 
 	@RequestMapping("/usr/article/getArticles")
@@ -55,7 +55,7 @@ public class UsrArticleController {
 	public ResultData<List<Article>> getArticles() {
 		List<Article> articles = articleService.getArticles();
 		
-		return ResultData.from("S-1","게시물 리스트",articles);
+		return ResultData.from("S-1","게시물 리스트","articles",articles);
 	}
 
 	@RequestMapping("/usr/article/doDelete")
@@ -65,28 +65,37 @@ public class UsrArticleController {
 		if(httpSession.getAttribute("loginedMemberId") ==null) {
 			return ResultData.from("F-A", "로그인 후 이용해주세요");
 		}
-		
+		// 아이디 받아와서 대조해서 맞으면 넘기는걸로 
+		int loginedMemberId = (int)httpSession.getAttribute("loginedMemberId");
 		Article article = articleService.getArticle(id);
 		if (article == null) {
 			return ResultData.from("F-1",Utility.f("%d 번 게시글은 존재하지 않습니다.", id));
 		}
+		if(loginedMemberId != article.getMemberId()) {
+			return ResultData.from("F-B","해당 게시물에 권한이 없습니다.");
+		}
 //		articles.remove(article);
 		articleService.deleteArticle(id);
-		return ResultData.from("S-1",Utility.f("%d 번 게시글을 삭제했습니다.", id),id);
+		return ResultData.from("S-1",Utility.f("%d 번 게시글을 삭제했습니다.", id),"id",id);
 	}
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public Object doModify(HttpSession httpSession , int id, String title, String body) {
+	public ResultData<Article> doModify(HttpSession httpSession , int id, String title, String body) {
 		if(httpSession.getAttribute("loginedMemberId") ==null) {
 			return ResultData.from("F-A", "로그인 후 이용해주세요");
 		}
+		int loginedMemberId = (int)httpSession.getAttribute("loginedMemberId");
 		Article article = articleService.getArticle(id);
 		if (article == null) {
-			return id + "번 게시글은 존재하지 않습니다.";
+			return ResultData.from("F-1",Utility.f("%d번 게시물은 존재하지 않습니다.", id));
 		}
-		articleService.modifyArticle(id, title, body);
-		return article;
+		ResultData<Article> actorCanModifyRd =articleService.actorCanModify(loginedMemberId,article);
+		if(actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
+		}
+		
+		return articleService.modifyArticle(id, title, body);
 	}
 	// 상세보기
 	@RequestMapping("/usr/article/getArticle")
@@ -98,6 +107,6 @@ public class UsrArticleController {
 			return ResultData.from("F-1",Utility.f("%d번 게시글은 존재하지 않습니다.", id));
 //			return id + "번 게시글은 존재하지 않습니다.";
 		}
-		return ResultData.from("S-1",Utility.f("%d번 게시글 입니다.", id),article);
+		return ResultData.from("S-1",Utility.f("%d번 게시글 입니다.", id),"article",article);
 	}
 }
